moved utilities for clauses into separate file
